Isolate Sanic app instances in tests
When adding tests, I found that trying to run the sanic app after a a
test module that already interacted with it would cause the app to fail
to bind for reasons I don't completely understand. As such, I isolated
app creation so tests can make an manage their own instances of the app,
which is a cleaner pattern than an app global anyway.