Cleanly shutdown CBS and queue watching thread
When running integration tests with the callback server via Callback, I
found that the tests would hang forever, I suspect because the queue
watching process would never terminate. This PR reworks shutdown for the
callback server and watch thread when setup via JobRunner.callback().
For now it doesn't alter how run() works, since that has watch loop and
CBS termination built in and seems to work as is.

The new stop() and wait_for_stop() methods will be exercised in the next
PR.

I manually tested that starting and curling the root of the CBS in a
container works.