Fix server death on module lookup fali
This is a bit of a hack but it's the best solution I can see for now
without a massive rewrite.

Currently if a module can't be found an exception will be thrown and the
JobRunner._watch loop will die, meaning the queues will not longer be
serviced. This checks the module lookup in the callback server before
submitting to any queues.

There's still another error that can occur if too many tasks are
submitted at the same time that also kills the watch loop.