Fix callback server death on task overage
If a caller attempts to run too many tasks on the callback server at
once when run standalone, the JobRunner._watch loop will exit and the
callback server will hang forever. This duplicates (for now) the job
counting logic in the callback server itself so it doesn't submit jobs
guaranteed to fail to the _watch loop. Eventually all this should likely
be refactored.

Also catches and reports bad module name errors.