Fix callback server dying when an error occurs in JobRunner._submit
Currently, when running in callback server mode, if the
JobRunner._submit mode throws an error, the _watch loop will exit and
the callback server will hang forever as the queues are no longer being
serviced. This handles any _submit errors by putting an error back on
the queue instead when running in callback server mode.