Port shifter changes from ee2 branch
Changes that are functionally the same as the changes in this commit are
already running on prod. Due to refactoring of how configuration works,
the changes on the ee2 branch could not be exactly replicated here, but
they should be functionally the same.

This is impossible to test other than in situ.